requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
qdrant-client==1.11.3
sentence-transformers==3.0.1
numpy==2.0.2
tqdm==4.66.5
rapidfuzz==3.9.7
requests==2.32.3
openai==1.40.3
